backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
msal==1.31.0
requests==2.32.3
PyPDF2==3.0.1
reportlab==4.0.7
xhtml2pdf==0.2.13
